sensor/bmi160.c:provides work in character device mode and UORB communication mode.
create a new bmi160.h file.
